+++ test/test_bases.jl
@@ -1,19 +1,19 @@
 using Test
-using QuantumInterface: tensor, ⊗, ptrace, reduced, permutesystems, multiplicable
+using QuantumInterface: dimension, shape, tensor, ⊗, ptrace, reduced, permutesystems, multiplicable
 using QuantumInterface: GenericBasis, CompositeBasis, NLevelBasis, FockBasis
 
 @testset "basis" begin
 
-shape1 = [5]
-shape2 = [2, 3]
-shape3 = [6]
+d1 = 5
+d2 = 2
+d3 = 6
 
-b1 = GenericBasis(shape1)
-b2 = GenericBasis(shape2)
-b3 = GenericBasis(shape3)
+b1 = GenericBasis(d1)
+b2 = GenericBasis(d2)
+b3 = GenericBasis(d3)
 
-@test b1.shape == shape1
-@test b2.shape == shape2
+@test dimension(b1) == d1
+@test dimension(b2) == d2
 @test b1 != b2
 @test b1 != FockBasis(2)
 @test b1 == b1
@@ -22,20 +22,20 @@ b3 = GenericBasis(shape3)
 comp_b1 = tensor(b1, b2)
 comp_uni = b1 ⊗ b2
 comp_b2 = tensor(b1, b1, b2)
-@test comp_b1.shape == [prod(shape1), prod(shape2)]
-@test comp_uni.shape == [prod(shape1), prod(shape2)]
-@test comp_b2.shape == [prod(shape1), prod(shape1), prod(shape2)]
+@test shape(comp_b1) == [d1, d2]
+@test shape(comp_uni) == [d1, d2]
+@test shape(comp_b2) == [d1, d1, d2]
 
 @test b1^3 == CompositeBasis(b1, b1, b1)
 @test (b1⊗b2)^2 == CompositeBasis(b1, b2, b1, b2)
 @test_throws DomainError b1^(0)
 
 comp_b1_b2 = tensor(comp_b1, comp_b2)
-@test comp_b1_b2.shape == [prod(shape1), prod(shape2), prod(shape1), prod(shape1), prod(shape2)]
+@test shape(comp_b1_b2) == [d1, d2, d1, d1, d2]
 @test comp_b1_b2 == CompositeBasis(b1, b2, b1, b1, b2)
 
 @test_throws ArgumentError tensor()
-@test comp_b2.shape == tensor(b1, comp_b1).shape
+@test shape(comp_b2) == shape(tensor(b1, comp_b1))
 @test comp_b2 == tensor(b1, comp_b1)
 
 @test_throws ArgumentError ptrace(comp_b1, [1, 2])
@@ -50,6 +50,6 @@ comp2 = tensor(b2, b1, b3)
 @test permutesystems(comp1, [2,1,3]) == comp2
 
 @test [b1, b2] != [b1, b3]
-@test !multiplicable(comp1, b1 ⊗ b2 ⊗ NLevelBasis(prod(b3.shape)))
+@test comp1 != b1 ⊗ b2 ⊗ NLevelBasis(prod(shape(b3)))
 
 end # testset